requirements.txt
aniso8601==6.0.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
enum34==1.1.6
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
functools32==3.2.3.post2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.1
MarkupSafe==1.1.1
psycopg2==2.7.7
PyJWT==1.7.1
pyrsistent==0.14.11
pytz==2018.9
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.0
urllib3==1.24.1
Werkzeug==0.14.1
